Added BalanceChange (number of BalanceChanges in Battlefield fits with Geth) and refactored heavily some part
This causes a lot of files to be touched because I had to change the signature of some
important trait like the `Ext` that describes external methods available to EVM and WASM
interpreter as well as going deeper passing deepmind::Tracer around.

Added deepmind::BlockTracer support for tracing stuff that happens outside of a transaction
in Ethereum (block mining rewards, DAO adjustements).

Eliminate dependency of deepmind on ethcore/vm and ethcore/types package as we started
having cycle in dependencies that required the heavy refactoring above. Now those packages
depends stricly on deep mind and converts they data to deep mind format.